[front/components/poke/templates] - refactor: consolidate template table components
 - Combine separate template table and columns components into a single TemplatesDataTable component
 - Remove redundant code and simplify the structure for displaying template data in tables
